Final cambios API (MVC)
