+++ api/v1/index.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once '/../includes/connection.php';
+require_once './../includes/connection.php';
 
 $method = $_SERVER['REQUEST_METHOD'];
 $uri = $_SERVER['REQUEST_URI'];
