+++ api/v1/modelos/delete-session.php
@@ -1,0 +1,9 @@
+<?php
+
+if(!isset($conn)) die();
+
+session_start();
+if(isset($_SESSION)){
+    unset($_SESSION);
+    session_destroy();
+}
